Cellular automata (multi-thread), tilemap, better trees
Added cellular automata: currently only tracks oxygen levels, but should be easy to extend to e.g. heat

Displays it as a sprite on the background.

Added tilemap and made player function with it.

Also, trees grow when you shoot them, and their collision testing is (mostly) how I want it.