Player animations, bug fixes, fancier home screen
Fixed a few bugs with bullets, wrong sound effects, not registering when player was in the air, etc.

Also added all the player animations (although only the "dying" animation is anything special)

And added an image to the home screen, to make it look nicer.

Now it's time to balance the game and create several stages, then I can submit it to the game jam!